drftemplate/requirements.txt
asgiref==3.8.1
Django==5.0.3
django-mssql==1.8
djangorestframework==3.15.1
pyodbc==5.1.0
python-dotenv==1.0.1
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
